requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.2
beautifulsoup4==4.12.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
filelock==3.10.7
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gspread==5.7.2
gspread-dataframe==3.3.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==2.0.0
parsel==1.7.0
pd==0.0.4
Protego==0.2.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
Scrapy==2.8.0
scrapy-splash==0.9.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4
tldextract==3.4.0
tonyg-rfc3339==0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
w3lib==2.1.1
yarl==1.8.2
zope.interface==6.0
